backend/requirements.txt
# YTShortMaker Python Backend Dependencies

# FastAPI server
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.17

# Whisper transcription
openai-whisper==20240930

# FFmpeg wrapper
ffmpeg-python==0.2.0

# Pydantic for data validation
pydantic==2.10.3

# CORS support
starlette==0.41.3

# Async file operations
aiofiles==24.1.0

# Google Generative AI for highlight detection
google-generativeai==0.8.3
